+++ cvat-ui/src/index.tsx
@@ -7,17 +7,18 @@ import ReactDOM from 'react-dom';
 import { connect, Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 
-import CVATApplication from './components/cvat-app';
+import CVATApplication from 'components/cvat-app';
 
-import createRootReducer from './reducers/root-reducer';
-import createCVATStore, { getCVATStore } from './cvat-store';
+import createRootReducer from 'reducers/root-reducer';
+import createCVATStore, { getCVATStore } from 'cvat-store';
+import logger, { LogType } from 'cvat-logger';
 
-import { authorizedAsync } from './actions/auth-actions';
-import { getFormatsAsync } from './actions/formats-actions';
-import { checkPluginsAsync } from './actions/plugins-actions';
-import { getUsersAsync } from './actions/users-actions';
-import { getAboutAsync } from './actions/about-actions';
-import { shortcutsActions } from './actions/shortcuts-actions';
+import { authorizedAsync } from 'actions/auth-actions';
+import { getFormatsAsync } from 'actions/formats-actions';
+import { checkPluginsAsync } from 'actions/plugins-actions';
+import { getUsersAsync } from 'actions/users-actions';
+import { getAboutAsync } from 'actions/about-actions';
+import { shortcutsActions } from 'actions/shortcuts-actions';
 import {
     resetErrors,
     resetMessages,
@@ -35,6 +36,7 @@ interface StateToProps {
     pluginsInitialized: boolean;
     pluginsFetching: boolean;
     userInitialized: boolean;
+    userFetching: boolean;
     usersInitialized: boolean;
     usersFetching: boolean;
     aboutInitialized: boolean;
@@ -68,6 +70,7 @@ function mapStateToProps(state: CombinedState): StateToProps {
 
     return {
         userInitialized: auth.initialized,
+        userFetching: auth.fetching,
         pluginsInitialized: plugins.initialized,
         pluginsFetching: plugins.fetching,
         usersInitialized: users.initialized,
@@ -112,3 +115,32 @@ ReactDOM.render(
     ),
     document.getElementById('root'),
 );
+
+window.onerror = (
+    message: Event | string,
+    source?: string,
+    lineno?: number,
+    colno?: number,
+    error?: Error,
+) => {
+    if (typeof (message) === 'string' && source && typeof (lineno) === 'number' && (typeof (colno) === 'number') && error) {
+        const logPayload = {
+            filename: source,
+            line: lineno,
+            message: error.message,
+            column: colno,
+            stack: error.stack,
+        };
+
+        const store = getCVATStore();
+        const state: CombinedState = store.getState();
+        const { pathname } = window.location;
+        const re = RegExp(/\/tasks\/[0-9]+\/jobs\/[0-9]+$/);
+        const { instance: job } = state.annotation.job;
+        if (re.test(pathname) && job) {
+            job.logger.log(LogType.sendException, logPayload);
+        } else {
+            logger.log(LogType.sendException, logPayload);
+        }
+    }
+};