+++ cvat-ui/src/reducers/about-reducer.ts
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
+import { boundariesActions, BoundariesActionTypes } from 'actions/boundaries-actions';
 import { AboutActions, AboutActionTypes } from 'actions/about-actions';
 import { AuthActions, AuthActionTypes } from 'actions/auth-actions';
 import { AboutState } from './interfaces';
@@ -23,7 +24,7 @@ const defaultState: AboutState = {
 
 export default function (
     state: AboutState = defaultState,
-    action: AboutActions | AuthActions,
+    action: AboutActions | AuthActions | boundariesActions,
 ): AboutState {
     switch (action.type) {
         case AboutActionTypes.GET_ABOUT: {
@@ -46,7 +47,8 @@ export default function (
                 fetching: false,
                 initialized: true,
             };
-        case AuthActionTypes.LOGOUT_SUCCESS: {
+        case AuthActionTypes.LOGOUT_SUCCESS:
+        case BoundariesActionTypes.RESET_AFTER_ERROR: {
             return {
                 ...defaultState,
             };