+++ cvat-ui/src/reducers/users-reducer.ts
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
+import { BoundariesActionTypes, boundariesActions } from 'actions/boundaries-actions';
 import { AuthActionTypes, AuthActions } from 'actions/auth-actions';
 import { UsersActionTypes, UsersActions } from 'actions/users-actions';
 import { UsersState } from './interfaces';
@@ -14,7 +15,7 @@ const defaultState: UsersState = {
 
 export default function (
     state: UsersState = defaultState,
-    action: UsersActions | AuthActions,
+    action: UsersActions | AuthActions | boundariesActions,
 ): UsersState {
     switch (action.type) {
         case UsersActionTypes.GET_USERS: {
@@ -37,10 +38,9 @@ export default function (
                 fetching: false,
                 initialized: true,
             };
+        case BoundariesActionTypes.RESET_AFTER_ERROR:
         case AuthActionTypes.LOGOUT_SUCCESS: {
-            return {
-                ...defaultState,
-            };
+            return { ...defaultState };
         }
         default:
             return state;